% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{Model Data}
\description{
R6 Class representing data that's external from resources and agents

It contains a single variable, 'events' for when the data holds a list of events
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{events}}{Any events that need to be tracked}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-model_data-new}{\code{model_data$new()}}
\item \href{#method-model_data-clone}{\code{model_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-model_data-new}{}}}
\subsection{Method \code{new()}}{
Creates a new model_data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_data$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new model data object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-model_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
