% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\docType{class}
\name{Simulation}
\alias{Simulation}
\title{Simulator}
\description{
Advances one or more villages through time
}
\section{Methods}{

\describe{
\item{\code{run_model()}}{Runs the simulation}
\item{\code{show_results(dependent_variable)}}{Displays the time dependant variables post-simulation}.
}
Creates a new Simulation instance
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{villages}}{A list of villages that the simulator will run}

\item{\code{length}}{The number of time steps in the simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Simulation$new()}}
\item \href{#method-run_model}{\code{Simulation$run_model()}}
\item \href{#method-show_results}{\code{Simulation$show_results()}}
\item \href{#method-clone}{\code{Simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(length = NA, villages = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{length}}{The number of years to run the simulation for}

\item{\code{villages}}{A list of villages that will be simulated
Advances each village a single time step}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_model"></a>}}
\if{latex}{\out{\hypertarget{method-run_model}{}}}
\subsection{Method \code{run_model()}}{
This this how the simulator 'runs' through the simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$run_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
Prints the plots of the data from each village in the simulator
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_results"></a>}}
\if{latex}{\out{\hypertarget{method-show_results}{}}}
\subsection{Method \code{show_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$show_results(dependent_variable = "population")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dependent_variable}}{The}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This should be done better; the plots look like garbage
}

\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
