% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/village_manager.R
\docType{class}
\name{village_manager}
\alias{village_manager}
\title{Village Manager}
\description{
This object manages all of the villages. It acts as an interface to them
}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Creates a new manager}
\item{\code{get_villages()}}{Gets all of the villages that the manager has}
\item{\code{get_village()}}{Retrieves a specific village from the manager, by name}
\item{\code{add_village()}}{Adds a village to the manager}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{villages}}{A list of village objects
Creates a new, village manager}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-village_manager-new}{\code{village_manager$new()}}
\item \href{#method-village_manager-get_villages}{\code{village_manager$get_villages()}}
\item \href{#method-village_manager-get_village}{\code{village_manager$get_village()}}
\item \href{#method-village_manager-add_resource}{\code{village_manager$add_resource()}}
\item \href{#method-village_manager-clone}{\code{village_manager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-village_manager-new"></a>}}
\if{latex}{\out{\hypertarget{method-village_manager-new}{}}}
\subsection{Method \code{new()}}{
Get a new instance of a village_manager
Gets all of the managed villages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{village_manager$new(villages)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-village_manager-get_villages"></a>}}
\if{latex}{\out{\hypertarget{method-village_manager-get_villages}{}}}
\subsection{Method \code{get_villages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{village_manager$get_villages()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of resources
Gets a village given a village name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-village_manager-get_village"></a>}}
\if{latex}{\out{\hypertarget{method-village_manager-get_village}{}}}
\subsection{Method \code{get_village()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{village_manager$get_village(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the requested village}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A village object
Adds a village to the manager.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-village_manager-add_resource"></a>}}
\if{latex}{\out{\hypertarget{method-village_manager-add_resource}{}}}
\subsection{Method \code{add_resource()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{village_manager$add_resource(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The villages to add}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-village_manager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-village_manager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{village_manager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
