% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VillageState.R
\docType{class}
\name{VillageState}
\alias{VillageState}
\title{Village State}
\description{
This is an object that represents the state of a village at a particular time.
}
\details{
This class acts as a type of record that holds the values of the different village variables. This class can be subclassed
to include more variables that aren't present.
}
\section{Methods}{



\describe{
\item{\code{propagate()}}{Advances the village a single time step}
\item{\code{as_tibble()}}{Turns the object into a tibble}.
}
Creates a new State
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{birthRate}}{The average birth rate of the village's citizens}

\item{\code{deathRate}}{The average death rate of the village's citizens}

\item{\code{carryingCapacity}}{The maximum number of villagers the village can sustain}

\item{\code{cropProductivity}}{Productivity for crops}

\item{\code{fishCatchRate}}{Rate of fish caught}

\item{\code{year}}{The year that the state represents}

\item{\code{population}}{The number of villagers in the village}

\item{\code{cropStock}}{The number of crops in the village}

\item{\code{fishStock}}{The number of fish in the village}

\item{\code{farmers}}{The number of farmers in the village}

\item{\code{fishers}}{The number of fishers in the village}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VillageState$new()}}
\item \href{#method-as_tibble}{\code{VillageState$as_tibble()}}
\item \href{#method-clone}{\code{VillageState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes all of the properties in the state to the ones passed in. This should
be called by subclasses during initialization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VillageState$new(
  birthRate = 0.085,
  deathRate = 0.07,
  carryingCapacity = 300,
  cropProductivity = 3,
  fishCatchRate = 2,
  year = 1,
  population = 100,
  cropStock = 300,
  fishStock = 200,
  farmers = 0,
  fishers = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{birthRate}}{The average birth rate of the village's citizens}

\item{\code{deathRate}}{The average death rate of the village's citizens}

\item{\code{carryingCapacity}}{The maximum number of villagers the village can sustain}

\item{\code{cropProductivity}}{Productivity for crops}

\item{\code{fishCatchRate}}{Rate of fish caught}

\item{\code{year}}{The year that the state represents}

\item{\code{population}}{The number of villagers in the village}

\item{\code{cropStock}}{The number of crops in the village}

\item{\code{fishStock}}{The number of fish in the village}

\item{\code{farmers}}{The number of farmers in the village}

\item{\code{fishers}}{The number of fishers in the village
Returns a tibble representation of the state}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When adding a new property, make sure to add it to the tibble
representation.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Sometimes it's useful to visualize the states. Tibbles are
the common data structure to hold the data. This method gives a tibble
with each property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VillageState$as_tibble()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a tibble representation of the state
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VillageState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
