% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\docType{class}
\name{simulation}
\alias{simulation}
\title{simulation}
\description{
Advances one or more villages through time
}
\section{Methods}{

\describe{
\item{\code{run_model()}}{Runs the simulation}
\item{\code{show_results(dependent_variable)}}{Displays the time dependant variables post-simulation}.
}
Creates a new Simulation instance
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{start_date}}{The Gregorian date that the simulation starts on}

\item{\code{end_date}}{The Gregorian date that the simulation ends on}

\item{\code{agents}}{All of the agents in the simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{simulation$new()}}
\item \href{#method-run_model}{\code{simulation$run_model()}}
\item \href{#method-clone}{\code{simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulation$new(start_date, end_date, agents)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{The date to start the simulation}

\item{\code{end_date}}{The date that the simulation should end}

\item{\code{agents}}{A list of agents that will be simulated
Runs the simulation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_model"></a>}}
\if{latex}{\out{\hypertarget{method-run_model}{}}}
\subsection{Method \code{run_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulation$run_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
